Actually download metadata in parallel
When using the gradle/gradle build as an example, we were never downloading metadata in parallel during graph resolution.
This is due to a number of things

1. Our UNCONSTRAINED executer was not actually unconstrained. Its worker threads still required a worker lease, and therefore
any work scheduled on an unconstrained executor would still have its concurrency limited to the number of cores on the machine.
We fix this by not requiring unconstrained workers to acquire worker leases.
2. When using filtered repositories, we aways returned MetadataFetchingCost.FAST if a dependency did not match our filter.
If one repository claims a fast fetching cost, then we short circuit and assume we do not need to download in parallel.
Presumably, we had one of the first repos in our build have a filter that did not match many dependencies, so for all metadata
fetching we would not download in parallel.

So, this commit properly unconstrains the UNCONSTRAINED executer, and fixes filtered repositories so they download metadata in parallel.

One thing to note is that we moved artifact resolution to the MAX_WORKERS executor instead of the UNCONSTRAINED executor, since we perform
transforms while doing artifact resolution. This means if UNCONSTRAINED actually worked the way it should have worked, we would have
been performing way more transforms in parallel than we intended.

In practice, since the unconstrained pool was already limited to the number of CPUs, this should make no difference. However, this
shows that we are not downloading artifacts with high parallelism. We should somehow update parallel artifact resolution to schedule
transforms on the MAX_WORKERS executor and file downloads on the UNCONSTRAINED executor.