Deprecate publishing dependencies on unpublished projects
Currently Gradle will let you publish a dependency on another project even if that other project
is not published. This will cause Gradle to publish a dependency with coordinates that do not make sense,
as the target project never gets published to the a remote repository.

We had to enforce this behavior manually in the Gradle build using https://github.com/gradle/gradle/pull/32459
however it would be better if this was done automatically for all Gradle users.

Without this verification, it is very easy to add dependencies to other projects that are not published, which
leads to broken published metadata